+++ mods/src/sound/repository.cpp
@@ -5,6 +5,7 @@
 #include <ctime>
 
 #include <libreborn/libreborn.h>
+#include <media-layer/audio.h>
 
 #include "sound.h"
 
@@ -314,6 +315,12 @@ static std::unordered_map<std::string, std::vector<std::string>> repository = {
             {
                 "PCM_fuse"
             }
+        },
+        {
+            "random.burp",
+            {
+                "PCM_burp"
+            }
         }
     }
 };
@@ -331,5 +338,18 @@ std::string _sound_pick(std::string sound) {
     } else {
         // Invalid Sound
         ERR("Invalid Sound: %s", sound.c_str());
+    }
+}
+
+// Resolve All Sounds
+void _sound_resolve_all() {
+    std::string source = _sound_get_source_file();
+    if (source.size() > 0) {
+        for (auto it : repository) {
+            for (std::string name : it.second) {
+                // Zero Volume Prevents An OpenAL Source From Being Allocated While Still Resolving The Sound
+                media_audio_play(source.c_str(), name.c_str(), 0.0f, 0.0f, 0.0f, 1.0f, 0.0f, 1.0f);
+            }
+        }
     }
 }